Переименовывает offer-card в offer-card-view
