+++ src/components/offer-card-view/offer-card.prop.js
